Add better /jobb routing
Use /jobb/id-abc instead of the shitty /jobbannonse?id=id-abc

The old url will 301 Moved Permanently to the new url

TODO implement proper slugs for joblistings to make better urls
